speechtoText-service/requirements.txt
functions-framework==3.0.0
google-cloud-speech==2.16.2
google-auth==2.6.2
google-cloud-storage==2.10.0